refactor: Global Navigation Bar 컴포넌트 파일 추가
